♻️ Remove cli-parameter and switch to config based parameters
- this removes several config-structs for sub-packages which are hard to maintain

+++ internal/mgtt/server/client.go
@@ -23,7 +23,7 @@ func (l *Listener) Accept() (newClient *client.MgttClient) {
 	newClient = &client.MgttClient{}
 
 	// init it
-	newClient.Init(newConnection, int64(config.Values.Timeout))
+	newClient.Init(newConnection, config.Values.Timeout)
 
 	return
 }
